backend/requirements.txt
djangorestframework
mysqlclient
django-cors-headers
djangorestframework-simplejwt
django-environ
django_extensions